Conflicts:
	profiles/coreos/arm64/package.provided

+++ sys-apps/systemd/systemd-9999.ebuild
@@ -1,17 +1,20 @@
 # Copyright 1999-2015 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/systemd/systemd-9999.ebuild,v 1.169 2015/05/30 13:58:45 floppym Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-apps/systemd/systemd-9999.ebuild,v 1.174 2015/06/27 18:00:47 floppym Exp $
 
 EAPI=5
 
+AUTOTOOLS_AUTORECONF=yes
+AUTOTOOLS_PRUNE_LIBTOOL_FILES=all
+PYTHON_COMPAT=( python{2_7,3_3,3_4} )
 CROS_WORKON_PROJECT="coreos/systemd"
 CROS_WORKON_REPO="git://github.com"
 
-if [[ "${PV}" == 9999 ]]; then
+if [[ ${PV} == 9999 ]]; then
 	# Use ~arch instead of empty keywords for compatibility with cros-workon
 	KEYWORDS="~amd64 ~arm64 ~arm ~x86"
 else
-	CROS_WORKON_COMMIT="9b174479806a66ff3a220a89291a38f8a4fed701"
+	CROS_WORKON_COMMIT="015325350548732458e61c193f5fab6f139f47fc"
 	KEYWORDS="amd64 arm64 ~arm ~x86"
 fi
 
@@ -20,9 +23,6 @@ fi
 # the later eclass's version to win. Only need src_unpack from workon.
 inherit cros-workon
 
-AUTOTOOLS_AUTORECONF=yes
-AUTOTOOLS_PRUNE_LIBTOOL_FILES=all
-PYTHON_COMPAT=( python{2_7,3_3,3_4} )
 inherit autotools-utils bash-completion-r1 linux-info multilib \
 	multilib-minimal pam python-single-r1 systemd toolchain-funcs udev \
 	user
@@ -32,18 +32,19 @@ HOMEPAGE="http://www.freedesktop.org/wiki/Software/systemd"
 
 LICENSE="GPL-2 LGPL-2.1 MIT public-domain"
 SLOT="0/2"
-IUSE="acl apparmor audit cryptsetup curl doc elfutils gcrypt gnuefi gudev http
-	idn +importd introspection kdbus +kmod +lz4 lzma nat pam policykit python
+IUSE="acl apparmor audit cryptsetup curl elfutils gcrypt gnuefi http
+	idn importd +kdbus +kmod +lz4 lzma nat pam policykit python
 	qrcode +seccomp selinux ssl sysv-utils terminal test vanilla xkb"
 
-# Gentoo removed the nls use flag, we'll keep it for now
-IUSE+=" nls symlink-usr"
+# CoreOS specific use flags
+IUSE+=" man symlink-usr"
 
-REQUIRED_USE="importd? ( curl gcrypt lzma )"
+REQUIRED_USE="importd? ( curl gcrypt lzma )
+	python? ( ${PYTHON_REQUIRED_USE} )"
 
 MINKV="3.8"
 
-COMMON_DEPEND=">=sys-apps/util-linux-2.25:0=
+COMMON_DEPEND=">=sys-apps/util-linux-2.26:0=
 	sys-libs/libcap:0=
 	!<sys-libs/glibc-2.16
 	acl? ( sys-apps/acl:0= )
@@ -53,7 +54,6 @@ COMMON_DEPEND=">=sys-apps/util-linux-2.25:0=
 	curl? ( net-misc/curl:0= )
 	elfutils? ( >=dev-libs/elfutils-0.158:0= )
 	gcrypt? ( >=dev-libs/libgcrypt-1.4.5:0=[${MULTILIB_USEDEP}] )
-	gudev? ( >=dev-libs/glib-2.34.3:2=[${MULTILIB_USEDEP}] )
 	http? (
 		>=net-libs/libmicrohttpd-0.9.33:0=
 		ssl? ( >=net-libs/gnutls-3.1.4:0= )
@@ -63,7 +63,6 @@ COMMON_DEPEND=">=sys-apps/util-linux-2.25:0=
 		app-arch/bzip2:0=
 		sys-libs/zlib:0=
 	)
-	introspection? ( >=dev-libs/gobject-introspection-1.31.1:0= )
 	kmod? ( >=sys-apps/kmod-15:0= )
 	lz4? ( >=app-arch/lz4-0_p119:0=[${MULTILIB_USEDEP}] )
 	lzma? ( >=app-arch/xz-utils-5.0.5-r1:0=[${MULTILIB_USEDEP}] )
@@ -107,20 +106,21 @@ DEPEND="${COMMON_DEPEND}
 	>=sys-kernel/linux-headers-${MINKV}
 	ia64? ( >=sys-kernel/linux-headers-3.9 )
 	virtual/pkgconfig
-	doc? ( >=dev-util/gtk-doc-1.18 )
 	gnuefi? ( >=sys-boot/gnu-efi-3.0.2 )
 	python? ( dev-python/lxml[${PYTHON_USEDEP}] )
 	terminal? ( media-fonts/unifont[utils(+)] )
 	test? ( >=sys-apps/dbus-1.6.8-r1:0 )"
 
 # Not required when building from unpatched tarballs, but we build from git.
-DEPEND="${DEPEND}
-	app-text/docbook-xml-dtd:4.2
-	app-text/docbook-xml-dtd:4.5
-	app-text/docbook-xsl-stylesheets
-	dev-libs/libxslt:0
-	dev-libs/gobject-introspection
+DEPEND+="
+	man? ( app-text/docbook-xml-dtd:4.2
+		app-text/docbook-xml-dtd:4.5
+		app-text/docbook-xsl-stylesheets
+		dev-libs/libxslt:0 )
+	terminal? ( ${PYTHON_DEPS} )
 	>=dev-libs/libgcrypt-1.4.5:0"
+
+REQUIRED_USE+=" terminal? ( ${PYTHON_REQUIRED_USE} )"
 
 pkg_pretend() {
 	local CONFIG_CHECK="~AUTOFS4_FS ~BLK_DEV_BSG ~CGROUPS
@@ -161,16 +161,10 @@ pkg_pretend() {
 }
 
 pkg_setup() {
-	use python && python-single-r1_pkg_setup
+	:
 }
 
 src_prepare() {
-	if use doc; then
-		gtkdocize --docdir docs/ || die
-	else
-		echo 'EXTRA_DIST =' > docs/gtk-doc.make
-	fi
-
 	# Bug 463376
 	sed -i -e 's/GROUP="dialout"/GROUP="uucp"/' rules/*.rules || die
 
@@ -182,6 +176,10 @@ src_configure() {
 	MY_UDEVDIR=$(get_udevdir)
 	# Fix systems broken by bug #509454.
 	[[ ${MY_UDEVDIR} ]] || MY_UDEVDIR=/lib/udev
+
+	if use python || use terminal; then
+		python_setup
+	fi
 
 	multilib-minimal_src_configure
 }
@@ -209,8 +207,6 @@ multilib_src_configure() {
 		# no deps
 		--enable-efi
 		--enable-ima
-		# used for stacktraces in log messages, leave off for now
-		--disable-elfutils
 
 		# Optional components/dependencies
 		$(multilib_native_use_enable acl)
@@ -218,24 +214,21 @@ multilib_src_configure() {
 		$(multilib_native_use_enable audit)
 		$(multilib_native_use_enable cryptsetup libcryptsetup)
 		$(multilib_native_use_enable curl libcurl)
-		$(multilib_native_use_enable doc gtk-doc)
 		$(multilib_native_use_enable elfutils)
 		$(use_enable gcrypt)
 		$(multilib_native_use_enable gnuefi)
-		$(use_enable gudev)
 		$(multilib_native_use_enable http microhttpd)
 		$(usex http $(multilib_native_use_enable ssl gnutls) --disable-gnutls)
 		$(multilib_native_use_enable idn libidn)
 		$(multilib_native_use_enable importd)
 		$(multilib_native_use_enable importd bzip2)
 		$(multilib_native_use_enable importd zlib)
-		$(multilib_native_use_enable introspection)
 		$(use_enable kdbus)
 		$(multilib_native_use_enable kmod)
 		$(use_enable lz4)
 		$(use_enable lzma xz)
+		$(multilib_native_use_enable man manpages)
 		$(multilib_native_use_enable nat libiptc)
-		$(use_enable nls)
 		$(multilib_native_use_enable pam)
 		$(multilib_native_use_enable policykit polkit)
 		$(multilib_native_use_with python)
@@ -248,9 +241,6 @@ multilib_src_configure() {
 		$(multilib_native_use_enable test dbus)
 		$(multilib_native_use_enable xkb xkbcommon)
 
-		# not supported (avoid automagic deps in the future)
-		--disable-chkconfig
-
 		# hardcode a few paths to spare some deps
 		QUOTAON=/usr/sbin/quotaon
 		QUOTACHECK=/usr/sbin/quotacheck
@@ -262,9 +252,13 @@ multilib_src_configure() {
 		--with-dbuspolicydir="${EPREFIX}/usr/share/dbus-1/system.d"
 		--with-dbussessionservicedir="${EPREFIX}/usr/share/dbus-1/services"
 		--with-dbussystemservicedir="${EPREFIX}/usr/share/dbus-1/system-services"
-		--with-dbusinterfacedir="${EPREFIX}/usr/share/dbus-1/interfaces"
 
 		--with-ntp-servers="0.coreos.pool.ntp.org 1.coreos.pool.ntp.org 2.coreos.pool.ntp.org 3.coreos.pool.ntp.org"
+
+		# The CoreOS epoch, Mon Jul  1 00:00:00 UTC 2013. Used by timesyncd
+		# as a sanity check for the minimum acceptable time. Explicitly set
+		# to avoid using the current build time.
+		--with-time-epoch=1372636800
 
 		# no default name servers
 		--with-dns-servers=
@@ -293,9 +287,6 @@ multilib_src_compile() {
 	if multilib_is_native_abi; then
 		emake "${mymakeopts[@]}"
 	else
-		# prerequisites for gudev
-		use gudev && emake src/gudev/gudev{enumtypes,marshal}.{c,h}
-
 		echo 'gentoo: $(BUILT_SOURCES)' | \
 		emake "${mymakeopts[@]}" -f Makefile -f - gentoo
 		echo 'gentoo: $(lib_LTLIBRARIES) $(pkgconfiglib_DATA)' | \
@@ -328,7 +319,6 @@ multilib_src_install() {
 			install-pkgconfiglibDATA
 			install-includeHEADERS
 			# safe to call unconditionally, 'installs' empty list
-			install-libgudev_includeHEADERS
 			install-pkgincludeHEADERS
 		)
 
@@ -355,7 +345,7 @@ multilib_src_install_all() {
 			dosym "${ROOTPREFIX-/usr}/bin/systemctl" ${prefix}/sbin/${app}
 		done
 		dosym "${ROOTPREFIX-/usr}/lib/systemd/systemd" ${prefix}/sbin/init
-	else
+	elif use man; then
 		# we just keep sysvinit tools, so no need for the mans
 		rm "${D}"/usr/share/man/man8/{halt,poweroff,reboot,runlevel,shutdown,telinit}.8 \
 			|| die
